Drop Game Link to download and view.
Gives the ability to drop a game link from the halite website, and it
will download and load the game, the game will be saved in /games
folder.

Probably not the best javascript ever as i only learnt yesteday,  also
the url to download the game includes a player id,  This is tsadmirals
but any seem to work, no login required.